made dashboard into a template & navbar into a component
